Split up object generation logic
No behavioral change, just moving code around.
